+++ Source/General/ApplicationSettings.vb
@@ -71,6 +71,7 @@ Public Class ApplicationSettings
     Public PreviewSize As Integer = 70
     Public ProcessPriority As ProcessPriorityClass = ProcessPriorityClass.BelowNormal
     Public ProgressHighlighting As Boolean = True
+    Public ProgressHighlightingColorName As String = "Orange"
     Public ProgressReformatting As Boolean = True
     Public ProjectsMruNum As Integer = 15
     Public RecentFramePositions As List(Of String)